feat: `Nullable<T>` lib type (Taylor's Version)
There is already a `NonNullable<T>`lib type, which does not seem specified in ECMA-262 and so should not be a valid basis for rejecting this PR, which I am sure many people will be thankful for as [a good chunk](https://www.google.com/search?q=%22Nullable%3CT%3E%22+includes%3Atypescript) of TS workspaces define this manually, usually as the `T | null` union.

This suggestion was resisted in #19944, presumably because it's hard to know whether to use `T | null` or the example provided under the [**old handbook**](https://www.typescriptlang.org/docs/handbook/advanced-types.html#mapped-types), `{ [P in keyof T]: T[P] | null }` (all properties on `T` assignable to `null`).

There is nothing that stops us from doing both, and constructing a utility type `Nullable<T>` which can be assigned either to `null` *or* `T` with all properties assignable to `null`:

```ts
type Nullable<T> = null | { [P in keyof T]: null | T[P] };
```

This is not an existing utility type and so it is not as if it's breaking. This would afford a lot of convenience without any apparent drawbacks as far as I can tell, and the other PR seems to have been closed mostly on the basis of vibes only, not for any real reason.

Ultimately it does seem kind of ridiculous to leave a `NonNullable<T>` type without defining a `Nullable<T>` one, and I suggest we take a stab at it.