Improve quickinfo baselines
Add a copy with of the file at the top with annotations, like
find-all-ref and rename baselines.

I'm not 100% sure about the formatting, but I don't think it's too
overbearing; for the few files with lots of tests, more decoration is
helpful.

I don't know whether sticking to jsonc syntax is worthwhile either. I
think the '// ' prefix hurts readability but might be worthwhile in some
future world where we need to query our baselines for information.