JSDoc: Remove a lot of checks in parser loop
The checks are now done in the scanner during scanning of jsdoc comment
text.

A couple of baselines change because the top-level @-detection for new
tags works the same way between top-level comment text and tag comment
text. This makes the parser detect more tags than before, specifically
in the case where non-whitespace precedes the first tag on the first
line. I think this is a fix since our parser tries to find tags anywhere
on the line, and previously didn't for the *first line only*.