Port pyright optimization that skips loop flow nodes that dont contain relevant references
Fixes #51525

This requied reworking how we compare references so we could equate them with a `set` - that change, in and of itself,
may actually also be perf positive (since we now get to cache the walk over the reference's AST), but both changes together
is somewhere between neutral and 4% gains in local testing, depending on the suite. I'll need to see if this bears out on
the much slower CI box, though, since 4% locally is < 0.1s, which is well in the realm of "OS scheduler variance" for some of these
tests on my box. Do note, unlike pyright, which in the linked issue's change, calculates the references used within a loop
during binding, we calculate it late on-demand and cache it in the checker. This is basically required for us, since we equate
references using the identity of the resolved symbol of the root node of the reference (which... may be a bit circular with
expression checking!).