scanNumber: smaller return value + side effects
Make scanNumber return only the SyntaxKind it scanned, and consistently
side-effect tokenValue to return the value. Previously, it sometimes
set tokenValue, and then returned a two-property object containing both
kind and tokenValue.

This might be more efficient for this particular case: returning objects
is a pattern I want to prevent spreading in the scanner.
So this change is mostly about style.

Notably, scanNumber already side-effects tokenValue in some places, so
this change makes the code more consistent.