fix(54760): Report error for 'declare type' followed by newline
The following TypeScript code does not report a diagnostic. Instead,
the TypeScript copmiler parses the code as if ASI was applied between
two tokens on the same line:

    var declare, type, T;

    declare /*ASI*/ type /*ASI*/
    T = null;

I think this behavior is confusing. TypeScript v4.3.5 used to report a
diagnostic for this code, which I think is better than the current
behavior of allowing this code without any diagnostic.

Instead of performing ASI between 'declare' and 'type', report a
diagnostic telling the user that a newline shouldn't be written after
'type'.