Updating EPC for Union type.
If the target type is an union, we'll check each type in the union.
If any of the types ar assignable without excess properties, then
that's the solution.

We'll find those properties, which are assignable in at least one
type and report those as causing the Object literal to be
unassignable.