Allow calls ending with unions of tuples
Previously, only calls with tuples were allowed, not unions of
tuples. For example, this already works:

```
f(string, [number, string], string) ==> f(string, number, string, string)
```

But this does not:

```
f(string, [string] | [number, number]) ==> (f, string, string | number, number | undefined)
```

This PR allows union types like these *as the last argument*. It's
possible to allow them anywhere, but quite a bit more complicated. And
the code is already complicated enough: getEffectiveCallArguments now
needs to return the minimum number of arguments as well as the argument
list (which is the maximum number of arguments).

Also, this transformation should not happen when the tuple union
argument is passed to a tuple union rest parameter: `[number] |
[string]` is assignable to `[number] | [string]` but the transformed
`number | string` is not. Checking for this requires passing
around even more information.

Bottom line: I'm not sure the complexity is worth the new code.
However, if this is a good idea, there are 3 things that need to be cleaned up:

1. More precise error messages--maybe. Right now error messages are
reported in terms of the maximum number of arguments:
2. Allow tuples with trailing `...T` types. Not done or tested, but
straightfoward.
3. Find a more elegant way to return the minimum number of arguments.

Fixes #42508
Supercedes #43882, but does less than half of that PR, in less than half
the code.