Do not support class expressions in IsolatedDeclarations.
Declaration emit for class expressions has some surprising
behaviors, such as printing 'any' in self-reference (#GH56479)
or non-consistent emit of private identifiers (#GH56145).

Replicating this behavior in IsolatedDeclaration doesn't seem
to be the goal that we want to achieve, thus we're going to
disallow this pattern for now until the issues around class
expressions are fixed.

Signed-off-by: Hana Joo <hanajoo@google.com>