Improve re-use of old program if imports changed but structure remains valid
Currently if a source file changes and new imports may be added for
similar existing modules, TypeScript will degrade from full program
re-use to only safe modules. This can be improved with a seemingly
small cost of having to sort and dedupe imports.

The use case is that in Angular we are optimizing for ideal program
re-use and ocasionally we will run transforms on user code that already
contain Angular imports. We may want to add additional Angular imports
but aren't bothering with re-using the actual existing import
declaration, so a new one is added. This new one now changes the length
of `file.imports` and therefore degrades the program re-use.

Angular could consider re-using the existing import, but that might be
difficult if it e.g. is a default import, or just technically
non-trivial without breaking exiting references etc. It seems more
natural if TypeScript would not consider this as a change in structure,
as clearly the module names. outgoing-edges in the graph remain the
same.

Mode overrides with imports may be an exception, but be aware that the
deduplication does not have an effect on the later re-use resolution
validation- in which case it would fail regardless. This is matching 1:1
with what would happen right now if an import mode changes. No new
behavior, except that a comment for this is now added.