Add missing members to array of objects in quickfix
If an object with missing properties is part of an array, then the token
where the error occurs is an OpenBraceToken, which was not handled by the
quickfix.

The correct type is available on the parent ObjectLiteralExpression.

Note that the ObjectLiteralExpression does not have a `text` property,
so place the `identifier` string for deduplication on the
InfoKind.ObjectLiteral structure.