feat: allow explicit ScriptKind extraFileExtensions
When configuring the language service host with extra file extensions,
the file extension is only regarded as supported if its `Deferred` or like
JS. If the host can delare a specific ScriptKind such as `TSX` for a file
then it will be stripped out with getSupportedExtensions.

This then later causes the programs rootFiles to not include the extra file
extension, and thus be excluded from the program.

If using `Deferred` then the file when parsed will be treated as TS and fail to
compile when using (J)TSX syntax. In majority of cases the script kind will be
known by the consumer and they can set the explicit script kind for the extra
file extension.

Adjust the logic for `getSupportedExtensions` to include any extra file extension
that are not built in. This way all explicit extra file extensions added
will be included in the project/programs root files and an explicit `scriptKind`
can be set.

resolves: #60125