Let isolated declarations fixer add toplevel imports more
The fixer has a tendency to create inline imports, whereas toplevel
imports are almost always more readable. And the code already exists,
it's just a matter of using it.

I assume there may be value in applying this in some other places, but
this is the one that's bugging me!

Fixes #60266.